Move helper functions `compute_position_of_point` and `compute_velocity_of_point` to `joint.py`
